# release-21.1: internal/batchskl: return error on index overflow